+++ src/remember.js
@@ -12,7 +12,7 @@ const port = process.env.EXPRESS_PORT;
 dotenv.config();
 
 const { MEMORIES_TABLE_NAME, MESSAGES_TABLE_NAME } = require("../config");
-const { supabase } = require("../helpers");
+const { supabase } = require('./supabaseclient.js');
 
 /**
  * Retrieves user memories from the database.
@@ -21,7 +21,7 @@ const { supabase } = require("../helpers");
  * @returns {Promise<Array>} - A promise that resolves to an array of user memories.
  */
 async function getUserMemory(userId, limit = 5) {
-  const { supabase } = require("../helpers");
+  const { supabase } = require('./supabaseclient.js');
   if (!userId) {
     logger.info("No userId provided to getUserMemory");
     return [];
@@ -59,7 +59,7 @@ async function getMemoriesBetweenDates(startDate, endDate) {
   logger.info(`Looking for memories between ${startDate} and ${endDate}`);
   logger.info(`Looking for memories between ${startDate.toISOString()} and ${endDate.toISOString()}`);
 
-  const { supabase } = require("../helpers");
+  const { supabase } = require('./supabaseclient.js');
   const response = await supabase
     .from(MEMORIES_TABLE_NAME)
     .select("*")
@@ -92,7 +92,7 @@ async function getMemoriesBetweenDates(startDate, endDate) {
  * @returns {Promise<Array<Object>>} - A promise that resolves to an array of memory objects.
  */
 async function getAllMemories(limit = 5) {
-  const { supabase } = require("../helpers");
+  const { supabase } = require('./supabaseclient.js');
   const { data, error } = await supabase
     .from(MEMORIES_TABLE_NAME)
     .select("*")
@@ -171,7 +171,7 @@ async function storeUserMemory(
   const [{ embedding }] = openAiEmbeddingResponse.data.data;
   logger.info(`Embedding length: ${embedding.length}`);
 
-  const { supabase } = require("../helpers");
+  const { supabase } = require('./supabaseclient.js');
   const { data, error } = await supabase
     // .from("storage")
     .from(MEMORIES_TABLE_NAME)
@@ -203,7 +203,7 @@ async function storeUserMemory(
  * @returns {Promise<Array<Object>>} - A promise that resolves to an array of memory objects.
  **/
 async function getResourceMemories(resourceId, limit = 5) {
-  const { supabase } = require("../helpers");
+  const { supabase } = require('./supabaseclient.js');
   const { data, error } = await supabase
     .from(MEMORIES_TABLE_NAME)
     .select("*")
@@ -226,7 +226,7 @@ async function getResourceMemories(resourceId, limit = 5) {
  * @returns {Promise<boolean>} - True if there is a memory of the file, false otherwise.
  */
 async function hasMemoryOfResource(resourceId) {
-  const { supabase } = require("../helpers");
+  const { supabase } = require('./supabaseclient.js');
   const { data, error } = await supabase
     .from(MEMORIES_TABLE_NAME)
     .select("created_at")
@@ -255,7 +255,7 @@ async function hasRecentMemoryOfResource(resourceId, recencyHours = 24) {
     return false;
   }
 
-  const { supabase } = require("../helpers");
+  const { supabase } = require('./supabaseclient.js');
 
   const { data, error } = await supabase
     .from(MEMORIES_TABLE_NAME)
@@ -294,7 +294,7 @@ async function hasRecentMemoryOfResource(resourceId, recencyHours = 24) {
  * @returns {Promise<object>} - A promise that resolves to the stored message data.
  */
 async function storeUserMessage({ username, channel, guild }, value) {
-  const { supabase } = require("../helpers");
+  const { supabase } = require('./supabaseclient.js');
   console.log("does supabase exist?", supabase);
   const { data, error } = await supabase
     // .from("messages")
@@ -478,7 +478,7 @@ async function memoryToEmbedding(memory) {
  * @returns {Promise<Array>} - A promise that resolves to an array of relevant memories.
  */
 async function getRelevantMemories(queryString, limit = 5) {
-  const { supabase } = require("../helpers");
+  const { supabase } = require('./supabaseclient.js');
   // make sure queryString is a string
   if (typeof queryString !== "string") {
     return logger.info("No query string provided to getRelevantMemories");
@@ -534,7 +534,7 @@ async function getRelevantMemories(queryString, limit = 5) {
  * @param {string} queryString - The query string to search for relevant memories.
  */
 async function getMemoriesByString(queryString) {
-  const { supabase } = require("../helpers");
+  const { supabase } = require('./supabaseclient.js');
   if (typeof queryString !== "string") {
     return logger.info("No query string provided to getRelevantMemories");
   }
